// Maskani Backend/خريطة/src/lib/api/types.ts
export interface DormData {
  dormID?: string;
  dormName: string;
  address: string;
  furnishedOrNot: boolean;
  distance: number;
  universityName: string;
  ownerName?: string;
  phone?: string;
  // universityID?: string;
  // ownerID: number;
}